Checking in streamlined code
